Move public asset route handlers under api/assets
