+++ backend/src/app/api/assets/public_free_assets.py
