+++ backend/src/app/api/assets/public_media_assets.py
